Fix upstream WINDOW_UPDATE handling.
This meant that if a lot of data was returned by the query (where a lot is some
tens of kB, not exactly sure of the number) then the query would take at least
5s. The 5s is the default period in which the background thread runs to give
more credit to the server, to get more data.
I have updated upstream libraries to handle WINDOW_UPDATEs as soon as data is
received, which is the correct thing to do, and makes for the least amount of
latency.
This addresses CB-969 and makes it much more manageable.
It might also help with CB-803 when a large number of items are in a block.