Disable sessions and sessions cookies.
One of the issues pointed out in the audit of the mobile wallet was that the
backend (the wallet-proxy) sends the session cookie even if the connection is
not secure and that this should be disabled. Since we do not use sessions for
anything it is simpler to just disable it entirely.